requirements.txt
flask
mysql-connector-python
PyMySQL
SQLAlchemy
psycopg2